perf: Improve CLS by setting logo dimensions and refactor component structure
- Set explicit width and height dimensions for `logo` to minimize layout shifts.
- Enhanced CLS by stabilizing `logo` placement.
- Refactored component file structure for better organization and maintainability.
