+++ src/components/common/local-image-loader.tsx
