+++ src/pages/root/_components/hero/social-buttons.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 
 import { SOCIALS } from "@/constants/collections";
-import { Hint } from "@/common/components/hint";
+import { Hint } from "@/components/common/hint";
 import { cn } from "@/lib/utils";
 
 interface SocialButtonsProps {
