+++ src/pages/root/layout.tsx
@@ -3,8 +3,8 @@ import { useEffect } from "react";
 import ReactGA from "react-ga4";
 
 import ToasterProvider from "@/lib/providers/toaster-provider";
-import { Header } from "@/common/components/header";
-import { Footer } from "@/common/components/footer";
+import { Header } from "@/components/header";
+import { Footer } from "@/components/footer";
 import { AppRoutes } from "@/routes/app-routes";
 
 const RootLayout = () => {
